Emscripten builds (untested)

+++ sdk/include/rckid/graphics/tile_grid.h
@@ -61,7 +61,7 @@ namespace rckid {
 
             TileInfo() = default;
             TileInfo(TileInfo const &) = default;
-            TileInfo(char c): tile_{c} {}
+            TileInfo(char c): tile_{static_cast<uint8_t>(c)} {}
 
             /** Sets tile info to given character.
